stores: Unify stores types into single SQLStore
