+++ internal/stores/contracts_test.go
@@ -1,30 +1,20 @@
 package stores
 
 import (
-	"encoding/hex"
 	"errors"
 	"fmt"
 	"reflect"
 	"testing"
 
-	"go.sia.tech/renterd/internal/consensus"
 	rhpv2 "go.sia.tech/renterd/rhp/v2"
 	"go.sia.tech/siad/crypto"
 	"go.sia.tech/siad/types"
 	"gorm.io/gorm/schema"
-	"lukechampine.com/frand"
 )
-
-// newTestSQLContractStore creates a new SQLContractStore for testing.
-func newTestSQLContractStore() (*SQLContractStore, error) {
-	dbName := hex.EncodeToString(frand.Bytes(32)) // random name for db
-	conn := NewEphemeralSQLiteConnection(dbName)
-	return NewSQLContractStore(conn, true)
-}
 
 // TestSQLContractStore tests SQLContractStore functionality.
 func TestSQLContractStore(t *testing.T) {
-	cs, err := newTestSQLContractStore()
+	cs, _, _, err := newTestSQLStore()
 	if err != nil {
 		t.Fatal(err)
 	}
@@ -150,49 +140,5 @@ func TestSQLContractStore(t *testing.T) {
 	}
 	if err := tableCountCheck(&dbValidSiacoinOutput{}, 0); err != nil {
 		t.Fatal(err)
-	}
-}
-
-// TestSQLHostSetStore tests the bus.HostSetStore methods on the SQLContractStore.
-func TestSQLHostSetStore(t *testing.T) {
-	cs, err := newTestSQLContractStore()
-	if err != nil {
-		t.Fatal(err)
-	}
-
-	// Check db before adding a set.
-	setName := "foo"
-	sets, err := cs.HostSets()
-	if err != nil {
-		t.Fatal(err)
-	}
-	if len(sets) != 0 {
-		t.Fatalf("expected 0 sets got %v", len(sets))
-	}
-	_, err = cs.HostSet(setName)
-	if !errors.Is(err, ErrHostSetNotFound) {
-		t.Fatal("should fail", err)
-	}
-
-	// Add another one.
-	pks := []consensus.PublicKey{{1, 2, 3}, {3, 2, 1}}
-	if err := cs.SetHostSet(setName, pks); err != nil {
-		t.Fatal(err)
-	}
-
-	// Check again.
-	sets, err = cs.HostSets()
-	if err != nil {
-		t.Fatal(err)
-	}
-	if len(sets) != 1 {
-		t.Fatalf("expected 0 sets got %v", len(sets))
-	}
-	set, err := cs.HostSet(setName)
-	if err != nil {
-		t.Fatal("should fail", err)
-	}
-	if !reflect.DeepEqual(set, pks) {
-		t.Fatal("set mismatch")
 	}
 }
