+++ internal/stores/hostdb.go
@@ -196,11 +196,6 @@ func NewJSONHostDB(dir string) (*JSONHostDB, modules.ConsensusChangeID, error) {
 const consensusInfoID = 1
 
 type (
-	// SQLHostDBStore is a helper type for interacting with a SQL-based HostDB.
-	SQLHostDBStore struct {
-		db *gorm.DB
-	}
-
 	// dbHost defines a hostdb.Interaction as persisted in the DB.
 	// Deleting a host from the db will cascade the deletion and also delete
 	// the corresponding announcements and interactions with that host.
@@ -295,52 +290,8 @@ func (i dbInteraction) convert() hostdb.Interaction {
 	}
 }
 
-// NewSQLHostDB uses a given Dialector to connect to a SQL database.  NOTE: Only
-// pass migrate=true for the first instance of SQLHostDB if you connect via the
-// same Dialector multiple times.
-func NewSQLHostDB(conn gorm.Dialector, migrate bool) (*SQLHostDBStore, modules.ConsensusChangeID, error) {
-	db, err := gorm.Open(conn, &gorm.Config{})
-	if err != nil {
-		return nil, modules.ConsensusChangeID{}, err
-	}
-
-	if migrate {
-		// Create the tables.
-		tables := []interface{}{
-			&dbHost{},
-			&dbInteraction{},
-			&dbAnnouncement{},
-			&dbConsensusInfo{},
-		}
-		if err := db.AutoMigrate(tables...); err != nil {
-			return nil, modules.ConsensusChangeID{}, err
-		}
-		if res := db.Exec("PRAGMA foreign_keys = ON", nil); res.Error != nil {
-			return nil, modules.ConsensusChangeID{}, res.Error
-		}
-	}
-
-	// Get latest consensus change ID or init db.
-	var ci dbConsensusInfo
-	err = db.Where(&dbConsensusInfo{ID: consensusInfoID}).
-		Attrs(dbConsensusInfo{
-			ID:   consensusInfoID,
-			CCID: modules.ConsensusChangeBeginning[:],
-		}).
-		FirstOrCreate(&ci).Error
-	if err != nil {
-		return nil, modules.ConsensusChangeID{}, err
-	}
-	var ccid modules.ConsensusChangeID
-	copy(ccid[:], ci.CCID)
-
-	return &SQLHostDBStore{
-		db: db,
-	}, ccid, nil
-}
-
 // Host returns information about a host.
-func (db *SQLHostDBStore) Host(hostKey consensus.PublicKey) (hostdb.Host, error) {
+func (db *SQLStore) Host(hostKey consensus.PublicKey) (hostdb.Host, error) {
 	var h dbHost
 	tx := db.db.Where(&dbHost{PublicKey: hostKey}).
 		Preload("Interactions").
@@ -354,7 +305,7 @@ func (db *SQLHostDBStore) Host(hostKey consensus.PublicKey) (hostdb.Host, error)
 
 // Hosts returns up to max hosts that have not been interacted with since
 // the specified time.
-func (db *SQLHostDBStore) Hosts(notSince time.Time, max int) ([]hostdb.Host, error) {
+func (db *SQLStore) Hosts(notSince time.Time, max int) ([]hostdb.Host, error) {
 	// Filter all hosts for the ones that have not been updated since a
 	// given time.
 	var foundHosts [][]byte
@@ -387,7 +338,7 @@ func (db *SQLHostDBStore) Hosts(notSince time.Time, max int) ([]hostdb.Host, err
 
 // RecordInteraction records an interaction with a host. If the host is not in
 // the store, a new entry is created for it.
-func (db *SQLHostDBStore) RecordInteraction(hostKey consensus.PublicKey, hi hostdb.Interaction) error {
+func (db *SQLStore) RecordInteraction(hostKey consensus.PublicKey, hi hostdb.Interaction) error {
 	return db.db.Transaction(func(tx *gorm.DB) error {
 		// Create a host if it doesn't exist yet.
 		if err := tx.FirstOrCreate(&dbHost{}, &dbHost{PublicKey: hostKey}).Error; err != nil {
@@ -405,7 +356,7 @@ func (db *SQLHostDBStore) RecordInteraction(hostKey consensus.PublicKey, hi host
 }
 
 // hosts returns all hosts int he db.
-func (db *SQLHostDBStore) hosts() ([]dbHost, error) {
+func (db *SQLStore) hosts() ([]dbHost, error) {
 	var hosts []dbHost
 	tx := db.db.Preload("Interactions").
 		Preload("Announcements").
@@ -417,7 +368,7 @@ func (db *SQLHostDBStore) hosts() ([]dbHost, error) {
 }
 
 // ProcessConsensusChange implements consensus.Subscriber.
-func (db *SQLHostDBStore) ProcessConsensusChange(cc modules.ConsensusChange) {
+func (db *SQLStore) ProcessConsensusChange(cc modules.ConsensusChange) {
 	height := cc.InitialHeight()
 	for range cc.RevertedBlocks {
 		height--
